Renamed, finding intersections improved, test cases added.
