+++ src/test/java/org/openstreetmap/atlas/checks/validation/intersections/BoundaryIntersectionCheckTest.java
@@ -1,0 +1,84 @@
+package org.openstreetmap.atlas.checks.validation.intersections;
+
+import org.junit.Assert;
+import org.junit.Rule;
+import org.junit.Test;
+import org.openstreetmap.atlas.checks.configuration.ConfigurationResolver;
+import org.openstreetmap.atlas.checks.validation.verifier.ConsumerBasedExpectedCheckVerifier;
+import org.openstreetmap.atlas.utilities.configuration.Configuration;
+
+/**
+ * @author srachanski
+ */
+public class BoundaryIntersectionCheckTest {
+    
+    
+//    private static final Logger logger = LoggerFactory.getLogger(BoundaryIntersectionsCheckTest.class);
+    
+    @Rule
+    public BoundaryIntersectionCheckTestRule setup = new BoundaryIntersectionCheckTestRule();
+    
+    @Rule
+    public ConsumerBasedExpectedCheckVerifier verifier = new ConsumerBasedExpectedCheckVerifier();
+    
+    private final Configuration configuration = ConfigurationResolver.emptyConfiguration();
+
+    @Test
+    public void testInvalidTwoCrossingItemsAtlas() {
+        this.verifier.actual(this.setup.crossingBoundariesTwoAreasIntersectEachOther(),
+                new BoundaryIntersectionCheck(this.configuration));
+        this.verifier.globallyVerify(flags -> Assert.assertEquals(2, flags.size()));
+        this.verifier.verify(flag -> Assert.assertEquals(5, flag.getFlaggedObjects().size()));
+    }
+    
+    @Test
+    public void testInvalidThreeCrossingItemsAtlas() {
+        this.verifier.actual(this.setup.crossingBoundariesTwoAreasIntersectOneOther(),
+                new BoundaryIntersectionCheck(this.configuration));
+        this.verifier.globallyVerify(flags -> Assert.assertEquals(3, flags.size()));
+    }
+    
+    @Test
+    public void testInvalidTwoCrossingItemsWithEdgesAtlas() {
+        this.verifier.actual(this.setup.crossingBoundariesTwoAreasIntersectEachOtherWithEdges(),
+                new BoundaryIntersectionCheck(this.configuration));
+        this.verifier.globallyVerify(flags -> Assert.assertEquals(2, flags.size()));
+        this.verifier.verify(flag -> Assert.assertEquals(5, flag.getFlaggedObjects().size()));
+    }
+    
+    @Test
+    public void testValidNonCrossingObjects() {
+        this.verifier.actual(this.setup.nonCrossingBoundariesTwoSeparate(),
+                new BoundaryIntersectionCheck(this.configuration));
+        this.verifier.globallyVerify(flags -> Assert.assertEquals(0, flags.size()));
+    }
+    
+    @Test
+    public void testValidNonCrossingObjectsOneContainOther() {
+        this.verifier.actual(this.setup.nonCrossingOneContainOther(),
+                new BoundaryIntersectionCheck(this.configuration));
+        this.verifier.globallyVerify(flags -> Assert.assertEquals(0, flags.size()));
+    }
+    
+    @Test
+    public void testValidNonCrossingObjectsWithEdges() {
+        this.verifier.actual(this.setup.nonCrossingBoundariesTwoSeparateWithEdges(),
+                new BoundaryIntersectionCheck(this.configuration));
+        this.verifier.globallyVerify(flags -> Assert.assertEquals(0, flags.size()));
+    }
+    
+    @Test
+    public void testValidCrossingObjectsOneMissingType() {
+        this.verifier.actual(this.setup.crossingOneWithWrongType(),
+                new BoundaryIntersectionCheck(this.configuration));
+        this.verifier.globallyVerify(flags -> Assert.assertEquals(0, flags.size()));
+    }
+    
+    @Test
+    public void testValidCrossingObjectsOneMissingBoundarySpecificTag() {
+        this.verifier.actual(this.setup.crossingOneMissingBoundarySpecificTag(),
+                new BoundaryIntersectionCheck(this.configuration));
+        this.verifier.globallyVerify(flags -> Assert.assertEquals(0, flags.size()));
+    }
+    
+}
